+++ dune/gdt/discretefunction/local.hh
@@ -146,13 +146,13 @@ public:
   typedef VectorImp                     VectorType;
   typedef typename BaseType::EntityType EntityType;
 
-  typedef typename BaseType::DomainFieldType  DomainFieldType;
-  static const unsigned int                   dimDomain = BaseType::dimDomain;
-  typedef typename BaseType::DomainType       DomainType;
+  typedef typename BaseType::DomainFieldType DomainFieldType;
+  static const size_t                        dimDomain = BaseType::dimDomain;
+  typedef typename BaseType::DomainType      DomainType;
 
   typedef typename BaseType::RangeFieldType RangeFieldType;
-  static const unsigned int                 dimRangeRows = BaseType::dimRangeCols;
-  static const unsigned int                 dimRangeCols = BaseType::dimRangeCols;
+  static const size_t                       dimRangeRows = BaseType::dimRangeCols;
+  static const size_t                       dimRangeCols = BaseType::dimRangeCols;
   typedef typename BaseType::RangeType      RangeType;
 
   typedef typename BaseType::JacobianRangeType JacobianRangeType;
@@ -241,13 +241,13 @@ public:
   typedef typename BaseType::VectorType VectorType;
   typedef typename BaseType::EntityType EntityType;
 
-  typedef typename BaseType::DomainFieldType  DomainFieldType;
-  static const unsigned int                   dimDomain = BaseType::dimDomain;
-  typedef typename BaseType::DomainType       DomainType;
+  typedef typename BaseType::DomainFieldType DomainFieldType;
+  static const size_t                        dimDomain = BaseType::dimDomain;
+  typedef typename BaseType::DomainType      DomainType;
 
   typedef typename BaseType::RangeFieldType RangeFieldType;
-  static const unsigned int                 dimRangeRows = BaseType::dimRangeCols;
-  static const unsigned int                 dimRangeCols = BaseType::dimRangeCols;
+  static const size_t                       dimRangeRows = BaseType::dimRangeCols;
+  static const size_t                       dimRangeCols = BaseType::dimRangeCols;
   typedef typename BaseType::RangeType      RangeType;
 
   typedef typename BaseType::JacobianRangeType JacobianRangeType;
